[Refactor] #40 - PrioritizationAPI 구조 개선 및 UI에서 임시 제거
